feat(test): 重整validator目录测试文件结构
将3个测试文件重组为1个标准命名的测试文件：
- 删除问题文件: locale_test.go, simple_coverage_test.go
- 重命名文件: validator_simple_test.go → validator_test.go
- 合并所有验证功能测试：结构验证、自定义验证器、多语言支持
- 保持66.5%测试覆盖率和29个测试函数
